Move SupportsBTF to bpfutils package
