+++ felix/bpf-gpl/calculate-flags
@@ -7,7 +7,7 @@
 filename=$1 # Example: from_wep_host_drop_fib_info.o
 args=()
 
-if [[ "${filename}" =~ .*core.* ]]; then
+if [[ "${filename}" =~ .*co-re.* ]]; then
   args+=("-DBPF_CORE_SUPPORTED")
 fi
 if [[ "${filename}" =~ .*debug.* ]]; then
