+++ felix/bpf/bpf.go
@@ -2240,15 +2240,6 @@ func KTimeNanos() int64 {
 	return ts.Nano()
 }
 
-func SupportsBTF() bool {
-	_, err := os.Stat("/sys/kernel/btf/vmlinux")
-	if err != nil {
-		log.WithError(err).Debug("BTF not supported")
-		return false
-	}
-	return true
-}
-
 const jumpMapVersion = 2
 
 func JumpMapName() string {
