+++ felix/bpf/tc/attach.go
@@ -38,6 +38,7 @@ import (
 	"github.com/projectcalico/calico/libcalico-go/lib/set"
 
 	"github.com/projectcalico/calico/felix/bpf"
+	"github.com/projectcalico/calico/felix/bpf/bpfutils"
 	"github.com/projectcalico/calico/felix/bpf/libbpf"
 )
 
@@ -59,7 +60,6 @@ type AttachPoint struct {
 	PSNATEnd             uint16
 	IPv6Enabled          bool
 	MapSizes             map[string]uint32
-	BTFEnabled           bool
 }
 
 var tcLock sync.RWMutex
@@ -122,7 +122,6 @@ func (ap AttachPoint) AttachProgram() (string, error) {
 		logCxt.WithError(err).Error("Failed to patch binary")
 		return "", err
 	}
-
 
 	// Using the RLock allows multiple attach calls to proceed in parallel unless
 	// CleanUpJumpMaps() (which takes the writer lock) is running.
@@ -368,7 +367,7 @@ func (ap *AttachPoint) ProgramID() (string, error) {
 
 // FileName return the file the AttachPoint will load the program from
 func (ap AttachPoint) FileName() string {
-	return ProgFilename(ap.Type, ap.ToOrFrom, ap.ToHostDrop, ap.FIB, ap.DSR, ap.LogLevel, ap.BTFEnabled)
+	return ProgFilename(ap.Type, ap.ToOrFrom, ap.ToHostDrop, ap.FIB, ap.DSR, ap.LogLevel, bpfutils.BTFEnabled)
 }
 
 func (ap AttachPoint) IsAttached() (bool, error) {
